modules/connection-api/requirements.txt
grpcio==1.51.3
protobuf==4.22.1
GeoAlchemy2==0.8.4
psycopg2-binary==2.8.5
shapely==1.7.0
SQLAlchemy==1.3.19